feature: 全局变量面板增加批量操作功能 #5140
