+++ frontend/desktop/src/assets/html/index-dev.html
@@ -8,7 +8,7 @@
     <title>标准运维</title>
     <script>
         var DEFAULT_PROJECT_ID = '2';
-        var SITE_URL = '/';
+        var SITE_URL = '/o/bk_sops/';
         var BK_STATIC_URL = ''
         var BK_DOC_URL = ''
         var FEEDBACK_URL = ''
