+++ frontend/desktop/src/config/i18n/cn.js
@@ -37,6 +37,7 @@ const cn = {
     '请填写变量值': '请填写变量值',
     '请填入项目变量说明': '请填入项目变量说明',
     '管理项目变量': '管理项目变量',
+    '变量快捷处理': '变量快捷处理',
     '运维工具': '运维工具',
     '监控告警': '监控告警',
     '配置管理': '配置管理',
@@ -1308,7 +1309,9 @@ const cn = {
     '变量配置': '变量配置',
     '变量复用': '变量复用',
     '手动创建': '手动创建',
-    '自动创建': '自动创建'
+    '自动创建': '自动创建',
+    '已选择x项': '已选择{num}项',
+    '确认删除所选的x个变量？': '确认删除所选的{num}个变量？'
 }
 
 export default cn
