+++ frontend/desktop/src/scss/iconfont.scss
@@ -497,12 +497,15 @@
 .common-icon-loading-round:before {
 	content: "\e1d0";
 }
-.common-icon-variable-cite:before {
-	content: "\e1d3";
-}
 .common-icon-render-skip:before {
 	content: "\e1d4";
 }
 .common-icon-perspective:before {
 	content: "\e1d5";
+}
+.common-icon-variable-cite:before {
+	content: "\e1d6";
+}
+.common-icon-drawable:before {
+	content: "\e1d7";
 }
